backend/ml_engine/pipelines/reconstruction.py
import os
from PIL import Image

class ReconstructionPipeline:
    def __init__(self):
        # 3D Reconstruction (PIFuHD / TripoSR) often requires external processes or heavier setup.
        # This wrapper prepares the input and calls the inference engine.
        pass

    def generate_3d(self, image_path: str) -> str:
        """
        Takes a 2D image path, runs 3D reconstruction, and returns the path to the .glb file.
        """
        # 3D Reconstruction requires a specific model (TripoSR/PIFuHD).
        # We raise error if not available rather than returning dummy data.
        raise NotImplementedError("3D Reconstruction Module not configured in this environment.")
        
        # In future:
        # output_path = ...
        # return output_path

# reconstruction_pipeline = ReconstructionPipeline()
