backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Caching
redis==5.0.1

# ML/AI - SDXL Body Generation (PRODUCTION)
torch==2.1.2
torchvision==0.16.2
diffusers==0.25.0
transformers==4.36.2
accelerate==0.25.0
safetensors==0.4.2
xformers==0.0.23

# Image Processing
pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
python-magic==0.4.27
python-magic-bin==0.4.14  # Windows binary for python-magic

# AI Services
google-generativeai==0.3.2

# Development
python-dotenv==1.0.0
einops==0.7.0
httpx==0.26.0
requests==2.31.0
slowapi==0.1.9
